+++ pages/auth/after-sign-in/google.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import AfterSignIn from '../../../components/after-sign-in';
+
+const AfterGoogleSignIn = () => <AfterSignIn mode="google" />;
+
+export default AfterGoogleSignIn;
