+++ pages/auth/after-sign-in/kakao.tsx
@@ -1,66 +1,6 @@
-import { useEffect } from 'react';
-import { NextPage } from 'next';
-import { useRouter } from 'next/router';
-import useSnackbarHandler from '../../../store/snackbar';
-import fetcherGenerator from '../../../fetcher-generator';
-import { TokenValidation } from '../../../index';
-import useAuth from '../../../store/auth';
-import Mine from '../../maps/mine';
+import React from 'react';
+import AfterSignIn from '../../../components/after-sign-in';
 
-const fetcher = fetcherGenerator<TokenValidation>();
-
-const AfterKakaoSignIn: NextPage = () => {
-    const { user, generateToken, validateToken, fetchUserInfo } = useAuth();
-    const { setMessage: setSnackbarMessage } = useSnackbarHandler();
-
-    const router = useRouter();
-
-    const onError = () => {
-        setSnackbarMessage(
-            'error',
-            '카카오 토큰 인증에서 오류가 발생했습니다.'
-        );
-
-        router.push('/');
-    };
-
-    // With code parameter
-    useEffect(() => {
-        if (user) {
-            if (!user.extraInfoInjected) {
-                router.push('/auth/extra-info');
-                return;
-            }
-            router.push('/');
-            return;
-        }
-
-        (async () => {
-            const params = new URL(document.location.toString()).searchParams;
-            const code: string | null = params.get('code');
-
-            if (!code) {
-                onError();
-
-                return;
-            }
-
-            const token = await generateToken(code);
-
-            console.log({ token });
-
-            const tokenStatus = token ? await validateToken() : false;
-
-            if (!tokenStatus) {
-                onError();
-                return;
-            }
-
-            await fetchUserInfo();
-        })();
-    }, [user]);
-
-    return <Mine />;
-};
+const AfterKakaoSignIn = () => <AfterSignIn mode="kakao" />;
 
 export default AfterKakaoSignIn;
